* getOppositeSolenoid() now receive and return int instead of uint8_t.

+++ Firmware.ino
@@ -245,7 +245,7 @@ uint8_t pinDigitalValueIs(int pin, int d){
 }
 
 // Receiving a pin, it returns the pin of the cylinder in charge of the opposite movement.
-uint8_t getOppositeSolenoid(uint8_t pinSol){
+int getOppositeSolenoid(int pinSol){
   
   uint8_t opposite=VALUE_PIN_NULL;
   if (pinSol==PIN_SOLU){opposite=PIN_SOLD;}
